Refactor code file name.
