+++ src/OpenArkDrv/driver-entry.cpp
@@ -16,10 +16,10 @@
 #include <ntifs.h>
 #include "arkdrv-api/arkdrv-api.h"
 #include "common/common.h"
-#include "driver/driver.h"
-#include "notify/notify.h"
-#include "memory/memory.h"
-#include "wingui/ops-hotkey/ops-hotkey.h"
+#include "kdriver/kdriver.h"
+#include "knotify/knotify.h"
+#include "kmemory/kmemory.h"
+#include "kwingui/ops-hotkey/ops-hotkey.h"
 #include "kstorage/kstorage.h"
 #include "kobject/kobject.h"
 #include "kprocess/kprocess.h"
